+++ montycat/__init__.py
@@ -2,3 +2,4 @@ from .core.engine import Engine
 from .core.tools import Pointer, Timestamp, Permission, PolicyCapability, PolicyKeyspaceType, SemanticModel, PolicyFormat
 from .core.schema import Schema
 from .core.store import Keyspace
+from .core.pool import PoolConfig, close_all_pools